fix: close semantic validation edge cases
